refactor(leaderboard): integrate EligibilityToggle into TopMinersTable and MinerRepoStandings
